requirements.txt
# Google Cloud Requirements
functions-framework>=3.0.0
gunicorn>=20.1.0
flask>=2.3.0
google-cloud-functions>=1.0.0
google-cloud-secret-manager>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
googleapis-common-protos>=1.0.0
google-api-python-client>=2.0.0

# Elaborazione dati
pandas>=1.5.0
openpyxl>=3.0.0

# HTTP requests
requests>=2.28.0

# Utilità
python-dotenv>=0.19.0